// src/api/generated/model/postApiV1CatCareCatsCatIdFluidInjectionsBodySite.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * parker-api
 * OpenAPI spec version: 0.1.0
 */

export type PostApiV1CatCareCatsCatIdFluidInjectionsBodySite = typeof PostApiV1CatCareCatsCatIdFluidInjectionsBodySite[keyof typeof PostApiV1CatCareCatsCatIdFluidInjectionsBodySite];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PostApiV1CatCareCatsCatIdFluidInjectionsBodySite = {
  left: 'left',
  right: 'right',
  nape: 'nape',
  other: 'other',
} as const;