// src/api/generated/model/postApiV1CatCareCatsCatIdFluidInjectionsBodyFluidType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * parker-api
 * OpenAPI spec version: 0.1.0
 */

export type PostApiV1CatCareCatsCatIdFluidInjectionsBodyFluidType = typeof PostApiV1CatCareCatsCatIdFluidInjectionsBodyFluidType[keyof typeof PostApiV1CatCareCatsCatIdFluidInjectionsBodyFluidType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PostApiV1CatCareCatsCatIdFluidInjectionsBodyFluidType = {
  normalSaline: 'normalSaline',
  lactatedRingers: 'lactatedRingers',
  other: 'other',
} as const;