// src/api/generated/model/getApiV1CatCareCatsCatIdFluidInjections200ItemSite.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * parker-api
 * OpenAPI spec version: 0.1.0
 */

export type GetApiV1CatCareCatsCatIdFluidInjections200ItemSite = typeof GetApiV1CatCareCatsCatIdFluidInjections200ItemSite[keyof typeof GetApiV1CatCareCatsCatIdFluidInjections200ItemSite];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiV1CatCareCatsCatIdFluidInjections200ItemSite = {
  left: 'left',
  right: 'right',
  nape: 'nape',
  other: 'other',
} as const;