// src/api/generated/model/getApiV1CatCareCatsCatIdFluidInjections200ItemFluidType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * parker-api
 * OpenAPI spec version: 0.1.0
 */

export type GetApiV1CatCareCatsCatIdFluidInjections200ItemFluidType = typeof GetApiV1CatCareCatsCatIdFluidInjections200ItemFluidType[keyof typeof GetApiV1CatCareCatsCatIdFluidInjections200ItemFluidType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiV1CatCareCatsCatIdFluidInjections200ItemFluidType = {
  normalSaline: 'normalSaline',
  lactatedRingers: 'lactatedRingers',
  other: 'other',
} as const;